навёл порядок в коде: выделил эффекты и шкалу в отдельные модули, починил отправку формы
